go 64bit, ≥X.5 only and UTF-8

+++ FlagChangeWindow.h
@@ -13,7 +13,7 @@
 	IBOutlet id		capsLockIndicatorTextField;
 }
 
-- (void) indicateCapsLock:(int) capsLock;
+- (void) indicateCapsLock:(NSInteger) capsLock;
 
 - (void)becomeKeyWindow;
